chore: extract pure functions from fiber/renderer.js
