+++ packages/react-devtools-shared/src/backend/types.js
@@ -101,7 +101,7 @@ export type FindHostInstancesForElementID = (
   id: number,
 ) => null | $ReadOnlyArray<HostInstance>;
 
-type Rect = {
+export type Rect = {
   x: number,
   y: number,
   width: number,
